refactor: DatabaseCleaner 패키지 분리
