+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.application;
 
+import kitchenpos.common.DatabaseCleaner;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
