+++ src/test/java/kitchenpos/common/DatabaseCleaner.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.common;
 
 import java.util.List;
 import java.util.stream.Collectors;
